+++ packages/terra-core-docs/src/terra-dev-site/doc/action-header/example/ActionHeaderNonUniqueIcon.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ActionHeader from 'terra-action-header';
 import ContentContainer from 'terra-content-container';
 import Card from 'terra-card';
-import IconPrinter from 'terra-icon/lib/icon/IconPrinter';
+import { IconPrinter } from 'terra-icon';
 import Button from 'terra-button';
 
 const ActionHeaderNonUniqueIcon = () => {