+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/example/CustomExample.jsx
@@ -1,19 +1,20 @@
 import React from 'react';
-import IconHelp from 'terra-icon/lib/icon/IconHelp';
 import classNames from 'classnames/bind';
+
 import Alert from 'terra-alert';
+import { IconHelp } from 'terra-icon';
 import styles from './colors.module.scss';
 
 const cx = classNames.bind(styles);
 
 const CustomExample = () => (
-  <Alert type="custom" title="Help!" customColorClass={cx(['my-app-alert-help-example'])} customIcon={<IconHelp />}>
-    <span>
-      This is a
-      <b> custom</b>
-      {' '}
-      notification banner
-    </span>
+  <Alert
+    type="custom"
+    title="Help!"
+    customColorClass={cx(['my-app-alert-help-example'])}
+    customIcon={<IconHelp />}
+  >
+    Welcome to Terra!
   </Alert>
 );