+++ packages/terra-list/src/ListSection.jsx
@@ -7,6 +7,7 @@ import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { injectIntl } from 'react-intl';
+import * as KeyCode from 'keycode-js';
 import SectionHeader from './ListSectionHeader';
 import styles from './List.module.scss';
 
@@ -82,7 +83,7 @@ const ListSection = ({
   ...customProps
 }) => {
   const [listItemNodes, setlistItemNodes] = useState(children);
-  let listNode = useRef();
+  let listSectionItemNode = useRef();
 
   useEffect(() => {
     if (!isCollapsible || !isCollapsed) {
@@ -101,8 +102,40 @@ const ListSection = ({
     cx('list', 'list-fill', theme.className),
   );
 
-  const handleListRef = (node) => {
-    listNode = node;
+  const handleListItemsRef = (nodes) => {
+    listSectionItemNode = nodes;
+  };
+
+  const handleKeyDown = event => {
+    const listItems = listSectionItemNode.querySelectorAll('[data-item-show-focus]');
+    const currentIndex = Array.from(listItems).indexOf(event.target);
+    const lastIndex = listItems.length - 1;
+
+    switch (event.nativeEvent.keyCode) {
+      case KeyCode.KEY_END:
+        event.preventDefault();
+        listItems[listItems.length - 1].focus();
+        break;
+      case KeyCode.KEY_HOME:
+        event.preventDefault();
+        listItems[0].focus();
+        break;
+      case KeyCode.KEY_UP: {
+        event.preventDefault();
+        const previousIndex = currentIndex > 0 ? currentIndex - 1 : lastIndex;
+        listItems[previousIndex].focus();
+        break;
+      }
+      case KeyCode.KEY_DOWN: {
+        event.preventDefault();
+        const nextIndex = currentIndex < lastIndex ? currentIndex + 1 : 0;
+        listItems[nextIndex].focus();
+        break;
+      }
+      default:
+        break;
+    }
+    event.stopPropagation();
   };
 
   const reorderListItems = (list, startIndex, endIndex) => {
@@ -124,8 +157,11 @@ const ListSection = ({
       result.destination.index,
     );
     setlistItemNodes(items);
-    if (listNode) {
-      listNode.focus();
+    if (listSectionItemNode) {
+      const listitems = listSectionItemNode.querySelectorAll('[data-item-show-focus]');
+      if (listitems && listitems[result.source.index]) {
+        listitems[result.source.index].focus();
+      }
     }
     provided.announce(intl.formatMessage({ id: 'Terra.list.drop' }, { startPosition: (result.source.index + 1), endPosition: (result.destination.index + 1) }));
     if (onDragEnd) {
@@ -137,18 +173,22 @@ const ListSection = ({
     isDraggable: ListItem?.props?.isSelectable,
     refCallback: (refobj) => {
       provider.innerRef(refobj);
-      handleListRef(refobj);
     },
     ...provider.draggableProps,
     ...provider.dragHandleProps,
   });
 
   const renderSectionListItemsDom = () => (
+    /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
     <>
       <SectionHeader {...customProps} isCollapsible={isCollapsible} isCollapsed={isCollapsed} />
       {listItemNodes && (
         <li className={cx('list-item')}>
-          <ul className={listClassNames}>
+          <ul
+            onKeyDown={handleKeyDown}
+            className={listClassNames}
+            ref={(refobj) => handleListItemsRef(refobj)}
+          >
             {listItemNodes}
           </ul>
         </li>
@@ -168,12 +208,22 @@ const ListSection = ({
 
   const renderDraggableListDom = () => (
     <DragDropContext onDragEnd={handleDragEnd} onDragStart={handleDragStart} onDragUpdate={handleDragUpdate}>
-      <Droppable droppableId="listSection">
+      <Droppable
+        droppableId="listSection"
+        renderClone={(provided, snapshot, rubric) => (
+          cloneListItem(listItemNodes[rubric.source.index], provided)
+        )}
+      >
         {(provided) => (
+          /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
           <div ref={provided.innerRef}>
             <SectionHeader {...customProps} isCollapsible={isCollapsible} isCollapsed={isCollapsed} />
             <li className={cx('list-item')}>
-              <ul className={listClassNames}>
+              <ul
+                ref={(refobj) => handleListItemsRef(refobj)}
+                className={listClassNames}
+                onKeyDown={handleKeyDown}
+              >
                 {listItemNodes.map((item, index) => (
                   <Draggable isDragDisabled={!(item?.props?.isSelectable)} key={item.key} draggableId={item.key} index={index}>
                     {(provider) => (