+++ packages/terra-list/src/ListSubsection.jsx
@@ -7,6 +7,7 @@ import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { injectIntl } from 'react-intl';
+import * as KeyCode from 'keycode-js';
 import SubsectionHeader from './ListSubsectionHeader';
 import styles from './List.module.scss';
 
@@ -79,7 +80,7 @@ const ListSubsection = ({
   ...customProps
 }) => {
   const [listItemNodes, setlistItemNodes] = useState(children);
-  let listNode = useRef();
+  let listSubSectionItemNode = useRef();
 
   useEffect(() => {
     if (!isCollapsible || !isCollapsed) {
@@ -98,9 +99,42 @@ const ListSubsection = ({
     cx('list', 'list-fill', theme.className),
   );
 
-  const handleListRef = (node) => {
-    listNode = node;
+  const handleListItemsRef = (nodes) => {
+    listSubSectionItemNode = nodes;
   };
+
+  const handleKeyDown = event => {
+    const listItems = listSubSectionItemNode.querySelectorAll('[data-item-show-focus]');
+    const currentIndex = Array.from(listItems).indexOf(event.target);
+    const lastIndex = listItems.length - 1;
+
+    switch (event.nativeEvent.keyCode) {
+      case KeyCode.KEY_END:
+        event.preventDefault();
+        listItems[listItems.length - 1].focus();
+        break;
+      case KeyCode.KEY_HOME:
+        event.preventDefault();
+        listItems[0].focus();
+        break;
+      case KeyCode.KEY_UP: {
+        event.preventDefault();
+        const previousIndex = currentIndex > 0 ? currentIndex - 1 : lastIndex;
+        listItems[previousIndex].focus();
+        break;
+      }
+      case KeyCode.KEY_DOWN: {
+        event.preventDefault();
+        const nextIndex = currentIndex < lastIndex ? currentIndex + 1 : 0;
+        listItems[nextIndex].focus();
+        break;
+      }
+      default:
+        break;
+    }
+    event.stopPropagation();
+  };
+
   const reorderListItems = (list, startIndex, endIndex) => {
     const result = Array.from(list);
     const [removed] = result.splice(startIndex, 1);
@@ -120,8 +154,11 @@ const ListSubsection = ({
       result.destination.index,
     );
     setlistItemNodes(items);
-    if (listNode) {
-      listNode.focus();
+    if (listSubSectionItemNode) {
+      const listitems = listSubSectionItemNode.querySelectorAll('[data-item-show-focus]');
+      if (listitems && listitems[result.source.index]) {
+        listitems[result.source.index].focus();
+      }
     }
     provided.announce(intl.formatMessage({ id: 'Terra.list.drop' }, { startPosition: (result.source.index + 1), endPosition: (result.destination.index + 1) }));
     if (onDragEnd) {
@@ -143,18 +180,18 @@ const ListSubsection = ({
     isDraggable: ListItem?.props?.isSelectable,
     refCallback: (refobj) => {
       provider.innerRef(refobj);
-      handleListRef(refobj);
     },
     ...provider.draggableProps,
     ...provider.dragHandleProps,
   });
 
   const renderSubSectionListItemsDom = () => (
+    /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
     <>
       <SubsectionHeader {...customProps} isCollapsible={isCollapsible} isCollapsed={isCollapsed} />
       {listItemNodes && (
       <li className={cx('list-item')}>
-        <ul className={listClassNames}>
+        <ul className={listClassNames} ref={(refobj) => handleListItemsRef(refobj)} onKeyDown={handleKeyDown}>
           {listItemNodes}
         </ul>
       </li>
@@ -164,12 +201,22 @@ const ListSubsection = ({
 
   const renderDraggableListDom = () => (
     <DragDropContext onDragEnd={handleDragEnd} onDragStart={handleDragStart} onDragUpdate={handleDragUpdate}>
-      <Droppable droppableId="listSubSection">
+      <Droppable
+        droppableId="listSubSection"
+        renderClone={(provided, snapshot, rubric) => (
+          cloneListItem(listItemNodes[rubric.source.index], provided)
+        )}
+      >
         {(provided) => (
+          /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
           <div ref={provided.innerRef}>
             <SubsectionHeader {...customProps} isCollapsible={isCollapsible} isCollapsed={isCollapsed} />
             <li className={cx('list-item')}>
-              <ul className={listClassNames}>
+              <ul
+                className={listClassNames}
+                onKeyDown={handleKeyDown}
+                ref={(refobj) => handleListItemsRef(refobj)}
+              >
                 {listItemNodes.map((item, index) => (
                   <Draggable isDragDisabled={!(item?.props?.isSelectable)} key={item.key} draggableId={item.key} index={index}>
                     {(provider) => (