+++ packages/terra-core-docs/CHANGELOG.md
@@ -7,18 +7,60 @@
   * Deleted examples of Terra Outline View Table as it is an obsolete package.
 
 * Updated
-  *  Updated doc site for `terra-form-single-select` and `terra-form-single-select-field`.
-  
-* Added
-  * Added drag and drop example for `terra-list`.
-  * Added instruction note for `terra-form-radio` label hidden example
+  * Updated `terra-alert` documentation for custom titles.
+
+## 1.44.1 - (October 11, 2023)
+
+* Fixed
+  * Fixed hyperlink button example import for `terra-hyperlink`.
+
+## 1.44.0 - (October 6, 2023)
 
 * Updated
-  *  Updated `terra-scroll` to add realistic examples.
-  *  Updated all paginator examples to use meaningful titles.
-  *  Updated testing recommendations to use `jest.spyOn` instead of `jest.mock`.
-  *  Updated `terra-alert` long text example for accessible `terra-show-hide`.
- 
+  * Updated test and doc examples for `terra-progress-bar`.
+
+## 1.43.0 - (October 3, 2023)
+
+* Added
+  * Added progressive paginator examples for custom page label and page count.
+
+* Updated
+  * Updated test example to validate `hidePageCount` prop.
+  * Updated terra-list examples to demonstrate arrow key navigation within list and list sections.
+
+## 1.42.0 - (September 26, 2023)
+
+* Updated
+  * Updated test drag and drop example for `terra-list`.
+
+* Added
+  * Added a test case for `terra-list` to support drag and drop in `terra-slide-panel`.
+
+## 1.41.0 - (September 21, 2023)
+
+* Fixed
+  * `terra-icon` import corrections in documentation and examples.
+  * Fixed formatting `terra-icon` accessibility guide.
+
+* Changed
+  * Updated icon example to specify `IconCeMarking` is private.
+
+## 1.40.0 - (September 19, 2023)
+
+* Added
+  * Added drag and drop example for `terra-list`.
+  * Added instruction note for `terra-form-radio` label hidden example.
+  * Added accessibility guidance for `terra-paginator`.
+
+* Updated
+  * Updated `terra-icon` imports in documentation.
+  * Updated `terra-scroll` to add realistic examples.
+  * Updated all paginator examples to use meaningful titles.
+  * Updated testing recommendations to use `jest.spyOn` instead of `jest.mock`.
+  * Updated `terra-alert` long text example for accessible `terra-show-hide`.
+  * Updated doc site for `terra-form-single-select` and `terra-form-single-select-field`.
+
+
 ## 1.39.0 - (August 25, 2023)
 
 * Fixed
@@ -114,7 +156,7 @@
 * Changed
   * Updated `terra-status-view - Change Variant` example component to use the apply button to change the selected variant.
   * Updated `terra-form-textarea` example component and added an button to toggle valid/invalid status
-  
+
 * Added
   * Added documentation updates for `Single Select` in `terra-form-select`.
 
@@ -179,7 +221,7 @@
 
 ## 1.24.0 - (April 5, 2023)
 
-* Added 
+* Added
   * Added Accessibility Guide for `terra-dropdown-button`.
 
 * Changed
@@ -220,7 +262,7 @@
 * Changed
   * Updated radio and checkbox tests to fix fusion theme tests.
 
-* Added 
+* Added
   * Added Accessibility Guide for `terra-toggle`.
   * Added A11y guide and examples for `terra-button`.
 
@@ -233,9 +275,9 @@
 * Changed
   * Updated examples to use informative icons.
 
-* Added 
+* Added
   * Added Label Example for `terra-toggle`.
-  
+
 ## 1.19.0 - (January 31, 2023)
 
 * Added