+++ packages/terra-dialog/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 2.80.0 - (October 3, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 2.79.0 - (September 21, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 2.78.0 - (September 19, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.77.0 - (August 25, 2023)
 
 * Changed