+++ packages/terra-search-field/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 3.93.0 - (October 3, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.92.0 - (September 21, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.91.0 - (September 19, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.90.0 - (August 25, 2023)
 
 * Changed